+++ core/scripts/run.sh
@@ -4,5 +4,5 @@ root_dirpath="$(dirname "${script_dirpath}")"
 
 GO_EXAMPLE_SUITE="kurtosistech/kurtosis-go-example"
 
-# TODO parameterize the binary name - it's hardcoded right now!
-"${root_dirpath}/build/kurtosis-core" "--test-suite-image=${GO_EXAMPLE_SUITE}" ${*}
+# The Go suite is designed to take in the nginxdemo/hello image - we only pass it in here as a demonstration of custom environment variables
+"${root_dirpath}/build/kurtosis-core" "--test-suite-image=${GO_EXAMPLE_SUITE}" '--custom-env-vars-json={"GO_EXAMPLE_SERVICE_IMAGE":"nginxdemos/hello"}' ${*}