+++ core/CHANGELOG.md
@@ -17,6 +17,7 @@
 * Made log levels of Kurtosis & test suite independently configurable
 * Switch to using CircleCI for builds
 * Made the API image & parallelism configurable
+* Allow configurable, custom Docker environment variables that will be passed as-is to the test suite
 
 # 0.9.0
 * Change ConfigurationID to be a string